Apply suggestion from code review
Remove IUserActionService and rename UserActionFactory to UserActionService
Add JsonPropertyName to classes returned through the ActionsController
