+++ src/Altinn.App.Api/Models/UserActionResponse.cs
@@ -1,7 +1,6 @@
 #nullable enable
-using System.Runtime.Serialization;
+using System.Text.Json.Serialization;
 using Altinn.App.Core.Models.UserAction;
-using Altinn.App.Core.Models.Validation;
 
 namespace Altinn.App.Api.Models;
 
@@ -13,15 +12,18 @@ public class UserActionResponse
     /// <summary>
     /// Data models that have been updated
     /// </summary>
+    [JsonPropertyName("updatedDataModels")]
     public Dictionary<string, object?>? UpdatedDataModels { get; set; }
     
     /// <summary>
     /// Actions frontend should perform after action has been performed backend
     /// </summary>
+    [JsonPropertyName("frontendActions")]
     public List<FrontendAction>? FrontendActions { get; set; }
     
     /// <summary>
     /// Validation issues that occured when processing action
     /// </summary>
+    [JsonPropertyName("error")]
     public ActionError? Error { get; set; }
 }
