+++ src/Altinn.App.Core/Extensions/ServiceCollectionExtensions.cs
@@ -248,8 +248,7 @@ namespace Altinn.App.Core.Extensions
 
         private static void AddActionServices(IServiceCollection services)
         {
-            services.TryAddTransient<UserActionFactory>();
-            services.TryAddTransient<IUserActionService, UserActionService>();
+            services.TryAddTransient<UserActionService>();
             services.AddTransient<IUserAction, NullUserAction>();
             services.AddTransient<IUserAction, SigningUserAction>();
             services.AddHttpClient<ISignClient, SignClient>();
