+++ src/Altinn.App.Core/Models/UserAction/ActionError.cs
@@ -1,3 +1,5 @@
+using System.Text.Json.Serialization;
+
 namespace Altinn.App.Core.Models.UserAction;
 
 /// <summary>
@@ -8,15 +10,18 @@ public class ActionError
     /// <summary>
     /// Machine readable error code
     /// </summary>
+    [JsonPropertyName("code")]
     public string Code { get; set; } = string.Empty;
 
     /// <summary>
     /// Human readable error message or text key
     /// </summary>
+    [JsonPropertyName("message")]
     public string Message { get; set; } = string.Empty;
     
     /// <summary>
     /// Error metadata
     /// </summary>
+    [JsonPropertyName("metadata")]
     public Dictionary<string, string>? Metadata { get; set; }
 }
